Migration
- Updated map fields to float

+++ db/migrate/20191126055926_add_location_to_user.rb
@@ -1,6 +1,6 @@
 class AddLocationToUser < ActiveRecord::Migration[5.2]
   def change
-    add_column :users, :longitude, :decimal
-    add_column :users, :latitude, :decimal
+    add_column :users, :longitude, :float
+    add_column :users, :latitude, :float
   end
 end
