+++ db/schema.rb
@@ -56,8 +56,8 @@ ActiveRecord::Schema.define(version: 2019_11_26_055926) do
     t.string "picture_content_type"
     t.integer "picture_file_size"
     t.datetime "picture_updated_at"
-    t.decimal "longitude", precision: 10
-    t.decimal "latitude", precision: 10
+    t.float "longitude"
+    t.float "latitude"
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
